--- a/pkg/mimirpb/compat_rw2.go
+++ b/pkg/mimirpb/compat_rw2.go
@@ -50,7 +50,7 @@ func (ps *rw2PagedSymbols) releasePages() {
 		*page = (*page)[:0]
 		rw2PagedSymbolsPool.Put(page)
 	}
-	ps.pages = ps.pages[:0]
+	ps.pages = nil
 	ps.count = 0
 }