--- a/pkg/mimirpb/mimir.pb.go
+++ b/pkg/mimirpb/mimir.pb.go
@@ -7584,6 +7584,7 @@ func (m *WriteRequest) Unmarshal(dAtA []byte) error {
 			m.Metadata = append(m.Metadata, metadata)
 		}
 		m.rw2symbols.releasePages()
+		m.unmarshalFromRW2 = false
 	}
 
 	return nil