@CLAassistant 2025-04-04
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/grafana/mimir?pullRequest=11100) <br/>Thank you for your submission! We really appreciate it. Like many open source projects, we ask that you all sign our [Contributor License Agreement](https://cla-assistant.io/grafana/mimir?pullRequest=11100) before we can accept your contribution.<br/>**17** out of **18** committers have signed the CLA.<br/><br/>:white_check_mark: dimitarvdimitrov<br/>:white_check_mark: lamida<br/>:white_check_mark: Ranexia04<br/>:white_check_mark: 56quarters<br/>:white_check_mark: ThePletch<br/>:white_check_mark: pracucci<br/>:white_check_mark: duricanikolic<br/>:white_check_mark: charleskorn<br/>:white_check_mark: colega<br/>:white_check_mark: pr00se<br/>:white_check_mark: seizethedave<br/>:white_check_mark: AndreZiviani<br/>:white_check_mark: krajorama<br/>:white_check_mark: codesome<br/>:white_check_mark: dmwilson-grafana<br/>:white_check_mark: aknuds1<br/>:white_check_mark: narqo<br/>:x: mimir-github-bot[bot]<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/grafana/mimir?pullRequest=11100) it.</sub>