@krajorama 2025-04-07
```
$ /home/krajo/opt/go/bin/go  test -count 5 -benchtime 5s -benchmem -run=^$ -tags requires_docker,stringlabels -bench ^BenchmarkUnMarshal$ github.com/grafana/mimir/pkg/mimirpb 
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/mimirpb
cpu: Intel(R) Core(TM) Ultra 7 155U
BenchmarkUnMarshal/RW1/skipExemplars=true-14         	     219	  27812540 ns/op	57122836 B/op	   60428 allocs/op
BenchmarkUnMarshal/RW1/skipExemplars=true-14         	     219	  26408683 ns/op	57122776 B/op	   60428 allocs/op
BenchmarkUnMarshal/RW1/skipExemplars=true-14         	     225	  26164794 ns/op	57122745 B/op	   60428 allocs/op
BenchmarkUnMarshal/RW1/skipExemplars=true-14         	     220	  26227564 ns/op	57122926 B/op	   60428 allocs/op
BenchmarkUnMarshal/RW1/skipExemplars=true-14         	     228	  26105894 ns/op	57122717 B/op	   60428 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=true-14         	     206	  29076566 ns/op	32797341 B/op	   50038 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=true-14         	     204	  27696413 ns/op	32797400 B/op	   50038 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=true-14         	     212	  28644989 ns/op	32797279 B/op	   50038 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=true-14         	     216	  28718101 ns/op	32801962 B/op	   50038 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=true-14         	     217	  28160463 ns/op	32797113 B/op	   50038 allocs/op
BenchmarkUnMarshal/RW1/skipExemplars=false-14        	     202	  29876341 ns/op	61922472 B/op	  100428 allocs/op
BenchmarkUnMarshal/RW1/skipExemplars=false-14        	     201	  30136930 ns/op	61922570 B/op	  100428 allocs/op
BenchmarkUnMarshal/RW1/skipExemplars=false-14        	     200	  31102272 ns/op	61922498 B/op	  100428 allocs/op
BenchmarkUnMarshal/RW1/skipExemplars=false-14        	     192	  30008941 ns/op	61922549 B/op	  100428 allocs/op
BenchmarkUnMarshal/RW1/skipExemplars=false-14        	     198	  29482934 ns/op	61922521 B/op	  100428 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=false-14        	     192	  30347598 ns/op	34403239 B/op	   60038 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=false-14        	     204	  29438473 ns/op	34397479 B/op	   60038 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=false-14        	     202	  29452409 ns/op	34397514 B/op	   60038 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=false-14        	     202	  30215303 ns/op	34407889 B/op	   60038 allocs/op
BenchmarkUnMarshal/RW2/skipExemplars=false-14        	     200	  31016646 ns/op	34392315 B/op	   60038 allocs/op
PASS
ok  	github.com/grafana/mimir/pkg/mimirpb	178.543s
```