@aknuds1 2025-04-07
@krajorama Could you compare your new benchmark stats with those from main (i.e. `benchstats`)?