@krajorama 2025-04-07
> @krajorama Could you compare your new benchmark results with those from main (i.e. `benchstats`)?

Main RW1 -> PR RW2
```
$ benchstat main-rw1.bench pr-rw2.bench 
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/mimirpb
cpu: Intel(R) Core(TM) Ultra 7 155U
                                    │ main-rw1.bench │            pr-rw2.bench             │
                                    │     sec/op     │    sec/op     vs base               │
UnMarshal/RW/skipExemplars=true-14       29.52m ± 2%   30.58m ± 29%       ~ (p=0.165 n=10)
UnMarshal/RW/skipExemplars=false-14      33.07m ± 2%   33.06m ±  3%       ~ (p=1.000 n=10)
geomean                                  31.25m        31.80m        +1.76%

                                    │ main-rw1.bench │             pr-rw2.bench             │
                                    │      B/op      │     B/op      vs base                │
UnMarshal/RW/skipExemplars=true-14      54.48Mi ± 0%   31.28Mi ± 0%  -42.58% (p=0.000 n=10)
UnMarshal/RW/skipExemplars=false-14     59.06Mi ± 0%   32.81Mi ± 0%  -44.45% (p=0.000 n=10)
geomean                                 56.72Mi        32.03Mi       -43.52%

                                    │ main-rw1.bench │            pr-rw2.bench             │
                                    │   allocs/op    │  allocs/op   vs base                │
UnMarshal/RW/skipExemplars=true-14       60.43k ± 0%   50.04k ± 0%  -17.19% (p=0.000 n=10)
UnMarshal/RW/skipExemplars=false-14     100.43k ± 0%   60.04k ± 0%  -40.22% (p=0.000 n=10)
geomean                                  77.90k        54.81k       -29.64%
```

Main RW1 -> PR RW1
```
benchstat main-rw1.bench pr-rw1.bench 
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/mimirpb
cpu: Intel(R) Core(TM) Ultra 7 155U
                                    │ main-rw1.bench │            pr-rw1.bench             │
                                    │     sec/op     │    sec/op     vs base               │
UnMarshal/RW/skipExemplars=true-14       29.52m ± 2%   29.37m ± 31%       ~ (p=0.684 n=10)
UnMarshal/RW/skipExemplars=false-14      33.07m ± 2%   33.37m ±  1%       ~ (p=0.063 n=10)
geomean                                  31.25m        31.31m        +0.20%

                                    │ main-rw1.bench │            pr-rw1.bench             │
                                    │      B/op      │     B/op      vs base               │
UnMarshal/RW/skipExemplars=true-14      54.48Mi ± 0%   54.48Mi ± 0%       ~ (p=0.493 n=10)
UnMarshal/RW/skipExemplars=false-14     59.06Mi ± 0%   59.06Mi ± 0%       ~ (p=1.000 n=10)
geomean                                 56.72Mi        56.72Mi       -0.00%

                                    │ main-rw1.bench │             pr-rw1.bench             │
                                    │   allocs/op    │  allocs/op   vs base                 │
UnMarshal/RW/skipExemplars=true-14       60.43k ± 0%   60.43k ± 0%       ~ (p=1.000 n=10)
UnMarshal/RW/skipExemplars=false-14      100.4k ± 0%   100.4k ± 0%       ~ (p=1.000 n=10) ¹
geomean                                  77.90k        77.90k       +0.00%
¹ all samples are equal
```